Added redshift adapter with tests, multiple commands causing failing
