+++ tests/integration/redshift_project/definitions/example_incremental_merge.sqlx
